User info route and admin privileges
Added /api/users/info route to retrieve either a list of all users or a single user and provided admin privileges for the routes /api/users/info, /api/users/logout and /api/users/delete.

+++ routes/api/users.js
@@ -7,6 +7,7 @@
  * @requires constants
  * @requires express
  * @requires jsonwebtoken
+ * @requires mongoose
  * @requires User
  * @requires utils
  * @requires validation
@@ -19,11 +20,11 @@ const bcryptjs = require("bcryptjs");
 const C = require("../../support/constants");
 const config = require("config");
 const jwt = require("jsonwebtoken");
+const mongoose = require("mongoose");
 const router = require("express").Router();
 const User = require("../../models/User");
 const utils = require("../../support/utilities");
 const validation = require("../../middleware/validation");
-
 /**
  * @description Returns an encrypted password.
  * 
@@ -89,65 +90,45 @@ const getJWTPayload = (userId) => {
 };
 
 /**
- * @description (GET) Get A list of all users.
- * Only admin users are authorized to retrieve the list of all user documents via token and admin authentication.
+ * @description Sends an error HTTP response according to a provided error message.
  * 
- * @protected
- * @constant
+ * @param {object} error - An error object that contains an error message.
+ * @param {object} response - An HTTP response object.
+ * @returns {object} An HTTP response object.
+ * @public
+ * @function 
  * 
  */
-router.get("/", auth, async (req, res) => {
-    
-    try {
 
-        const user = res.locals.user;
+const sendErrorResponse = (error, response) => {
 
-        if (user.admin) {
+    switch (error.message) {
 
-            const users = await User
-                .find({})
-                .select(`${C.Model.USER_NAME} ${C.Model.USER_EMAIL}`);
+        case C.Error.USER_ALREADY_EXISTS:
 
-            return res
-                .status(C.Status.OK)
-                .json({ users });
-        }
-        else {
+             return response
+                .status(C.Status.BAD_REQUEST)
+                .send({ error: C.Error.USER_ALREADY_EXISTS });
 
-            throw new Error(C.Error.USER_INVALID_CREDENTIALS);
-        }
-    }
-    catch (error) {
-        
-        if (error.message === C.Error.USER_INVALID_CREDENTIALS) {
+        case C.Error.USER_DOES_NOT_EXIST:
 
-            return res
+            return response
+                .status(C.Status.BAD_REQUEST)
+                .send({ error: C.Error.USER_DOES_NOT_EXIST });
+
+        case C.Error.USER_INVALID_CREDENTIALS:
+
+            return response
                 .status(C.Status.UNAUTHENTICATED)
-                .send(C.Error.USER_INVALID_CREDENTIALS);
-        }
+                .send({ error: C.Error.USER_INVALID_CREDENTIALS });
 
-        return res
-            .status(C.Status.INTERNAL_SERVER_ERROR)
-            .send(error.message);
+        default:
+
+            return response
+                .status(C.Status.INTERNAL_SERVER_ERROR)
+                .send({ error: error.message });
     }
-});
-
-/**
- * @description (GET) Check user authentication.
- * User authentication is verified by the presence and/or validity of the token value in the user document.
- * 
- * @protected
- * @constant
- * 
- */
-router.get(C.Route.AUTH, auth, (req, res) => {
-    
-    const user = res.locals.user;
-
-    return res
-        .status(C.Status.OK)
-        .json({ user });
-});
+};
 
 /**
  * @description (POST) Register user with input validation and password encryption.
@@ -204,16 +185,7 @@ router.post(C.Route.USERS_REGISTER, [
         }
         catch (error) {
 
-            if (error.message === C.Error.USER_ALREADY_EXISTS) {
-
-                return res
-                    .status(C.Status.BAD_REQUEST)
-                    .send(C.Error.USER_ALREADY_EXISTS);
-            }
-
-            return res
-                .status(C.Status.INTERNAL_SERVER_ERROR)
-                .send(error.message);
+            sendErrorResponse(error, res);
         }
     }
 );
@@ -267,19 +239,76 @@ router.post(C.Route.USERS_LOGIN, [
         }
         catch (error) {
 
-            if (error.message === C.Error.USER_INVALID_CREDENTIALS) {
-
-                return res
-                .status(C.Status.UNAUTHENTICATED)
-                .send(C.Error.USER_INVALID_CREDENTIALS);
-            }
-
-            return res
-                .status(C.Status.INTERNAL_SERVER_ERROR)
-                .send(error.message);
+            sendErrorResponse(error, res);
         }
     }
 );
+
+/**
+ * @description (GET) Information on one of more users.
+ * All users are authorized to retrieve their own user document via token authentication and optionally providing their own valid user ID as a request parameter.
+ * Admin users, via admin authentication, are authorized to retrieve either a list of all users or a single user by optionally providing a valid user ID as a request parameter.
+ * 
+ * @protected
+ * @constant
+ * 
+ */
+router.get(`${C.Route.USERS_INFO}/:${C.Route.PARAM_USER_ID}?`, auth, async (req, res) => {
+    
+    try {
+
+        const user = res.locals.user;
+        const paramUserID = req.params[C.Route.PARAM_USER_ID];
+        let result;
+
+        if (user.admin) {
+
+            if (paramUserID) {
+
+                const isValidUserID = mongoose.Types.ObjectId.isValid(paramUserID);
+
+                if (!isValidUserID) {
+
+                    throw new Error(C.Error.USER_DOES_NOT_EXIST);
+                }
+
+                result = (paramUserID === user.id) ? user : await User
+                    .findById(paramUserID)
+                    .select(`${C.Model.USER_NAME} ${C.Model.USER_EMAIL} ${C.Model.USER_ADMIN}`);
+
+                if (!result) {
+
+                    throw new Error(C.Error.USER_DOES_NOT_EXIST);
+                }
+
+                result = [result];
+            }
+            else {
+
+                result = await User
+                    .find({})
+                    .select(`${C.Model.USER_NAME} ${C.Model.USER_EMAIL} ${C.Model.USER_ADMIN}`);
+            }
+        }
+        else {
+
+            if (paramUserID && paramUserID !== user.id) {
+                
+                throw new Error(C.Error.USER_INVALID_CREDENTIALS);
+            }
+
+            result = [user];
+        }
+            
+        return res
+            .status(C.Status.OK)
+            .send({ info: result });
+    }
+    catch (error) {
+
+        sendErrorResponse(error, res);
+    }
+});
 
 /**
  * @description (PATCH) Update a user.
@@ -336,46 +365,81 @@ router.patch(C.Route.USERS_UPDATE, [
         }
         catch (error) {
 
-            return res
-                .status(C.Status.INTERNAL_SERVER_ERROR)
-                .send(error.message);
+            sendErrorResponse(error, res);
         }
     }
 );
 
 /**
- * @description (GET) Logout a user.
- * All users are authorized to logout and revoke authorization by deleting the token property from their own user document. 
+ * @description (GET) Information on one of more users.
+ * All users are authorized to log out via token authentication and optionally providing their own valid user ID as a request parameter.
+ * Admin users, via admin authentication, are authorized to log out any user by providing an optional valid user ID request parameter.
  * 
  * @protected
  * @constant
  * 
  */
-router.get(C.Route.USERS_LOGOUT, auth, async (req, res) => {
-
-    const user = res.locals.user;
-
+router.get(`${C.Route.USERS_LOGOUT}/:${C.Route.PARAM_USER_ID}?`, auth, async (req, res) => {
+    
     try {
-
-        user[C.Model.USER_TOKEN] = undefined;
         
-        await user.save();
+        const user = res.locals.user;
+        const paramUserID = req.params[C.Route.PARAM_USER_ID];
+        let result;
+
+        if (user.admin) {
+
+            if (paramUserID) {
+
+                const isValidUserID = mongoose.Types.ObjectId.isValid(paramUserID);
+
+                if (!isValidUserID) {
+
+                    throw new Error(C.Error.USER_DOES_NOT_EXIST);
+                }
+
+                result = (paramUserID === user.id) ? user : await User
+                    .findById(paramUserID)
+                    .select(`${C.Model.USER_NAME} ${C.Model.USER_EMAIL} ${C.Model.USER_ADMIN}`);
+
+                if (!result) {
+
+                    throw new Error(C.Error.USER_DOES_NOT_EXIST);
+                }
+            }
+            else {
+
+                result = user;
+            }
+        }
+        else {
+
+            if (paramUserID && paramUserID !== user.id) {
+                
+                throw new Error(C.Error.USER_INVALID_CREDENTIALS);
+            }
+
+            result = user;
+        }
+
+        result[C.Model.USER_TOKEN] = undefined;
+        
+        await result.save();
 
         return res
             .status(C.Status.OK)
-            .json({ user });
+            .send({ user: result });
     }
     catch (error) {
 
-        return res
-            .status(C.Status.INTERNAL_SERVER_ERROR)
-            .send(error.message);
+        sendErrorResponse(error, res);
     }
 });
 
 /**
  * @description (DELETE) Delete a user document.
- * All users are authorized to delete their own user document via token authentication.
+ * 
+ * All users are authorized to delete their own user document via token authentication and optionally providing their own valid user ID as a request parameter.
  * Admin users, via admin authentication, are authorized to delete any user document by providing an optional valid user ID request parameter.
  * 
  * @protected
@@ -388,47 +452,50 @@ router.delete(`${C.Route.USERS_DELETE}/:${C.Route.PARAM_USER_ID}?`, auth, async 
         
         const user = res.locals.user;
         const paramUserID = req.params[C.Route.PARAM_USER_ID];
-        let deletedUser;
+        let result;
 
-        if (paramUserID) {
+        if (user.admin) {
 
-            if (user.admin) {
+            if (paramUserID) {
 
-                deletedUser = await User
+                const isValidUserID = mongoose.Types.ObjectId.isValid(paramUserID);
+
+                if (!isValidUserID) {
+
+                    throw new Error(C.Error.USER_DOES_NOT_EXIST);
+                }
+
+                result = (paramUserID === user.id) ? user : await User
                     .findByIdAndRemove(paramUserID)
-                    .select(`${C.Model.USER_NAME} ${C.Model.USER_EMAIL}`);
+                    .select(`${C.Model.USER_NAME} ${C.Model.USER_EMAIL} ${C.Model.USER_ADMIN}`);
 
-                if (!deletedUser) {
+                if (!result) {
                     
-                    throw new Error(C.Error.USER_INVALID_CREDENTIALS);
+                    throw new Error(C.Error.USER_DOES_NOT_EXIST);
                 }
             }
             else {
 
-                throw new Error(C.Error.USER_INVALID_CREDENTIALS);
+                result = await user.deleteOne();
             }
         }
         else {
 
-            deletedUser = await user.deleteOne();
+            if (paramUserID && paramUserID !== user.id) {
+
+                throw new Error(C.Error.USER_INVALID_CREDENTIALS);
+            }
+        
+            result = await user.deleteOne();
         }
 
         return res
             .status(C.Status.OK)
-            .json({ deletedUser });
+            .json({ user: result });
     }
     catch (error) {
-        
-        if (error.message === C.Error.USER_INVALID_CREDENTIALS) {
 
-            return res
-                .status(C.Status.UNAUTHENTICATED)
-                .send(C.Error.USER_INVALID_CREDENTIALS);
-        }
-
-        return res
-            .status(C.Status.INTERNAL_SERVER_ERROR)
-            .send(error.message);
+        sendErrorResponse(error, res);
     }
 });
